Split basectl rpc into per-service modules
