+++ crates/infra/basectl/src/rpc/p2p.rs
@@ -1,0 +1,1 @@
+//! P2P RPC helpers will land with the non-TUI p2p command group.
